pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "mayaku"
version = "1.0.0"
description = "Backend-portable detection / segmentation / keypoint library inspired by Detectron2."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    {name = "Emrah Nazif", email = "emrah@datamarkin.com"},
    {name = "Datamarkin", email = "support@datamarkin.com"}
]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS :: MacOS X",
]
dependencies = [
  # Pinned per BACKEND_PORTABILITY_REPORT Appendix A.
  "torch>=2.4",
  "torchvision>=0.19",
  "numpy>=1.26",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "typer>=0.12",
  "pycocotools>=2.0.7",
  "pillow>=10.0",
]

[project.optional-dependencies]
# onnxruntime is intentionally not pinned here: CPU and CUDA hosts need
# different wheels (onnxruntime vs onnxruntime-gpu). Install manually:
#   CPU / Apple Silicon:  pip install onnxruntime
#   CUDA Linux/Windows:   pip install onnxruntime-gpu
# The CLI reports the right install command if the package is missing.
onnx = ["onnx>=1.16"]
coreml = ["coremltools>=7.2"]
openvino = ["openvino>=2024.3"]
# TensorRT is CUDA-only. NVIDIA ships wheels for Linux x86_64,
# Linux aarch64, and Windows x86_64; macOS has no NVIDIA GPU
# support so the upstream `tensorrt` package's download stub fails
# at install time on Darwin. The PEP 508 marker makes the dep
# silently no-op on macOS so a single
# `pip install -e '.[dev,onnx,coreml,openvino,tensorrt]'` line
# works on every host without conditional logic from the user.
# `onnx` is listed here so TRT export can reuse the ONNX front-end
# without requiring the user to also install `[onnx]` separately.
tensorrt = [
  "onnx>=1.16",
  "tensorrt>=10.0; sys_platform != 'darwin'",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-PyYAML",
  "types-Pillow",
]

[project.scripts]
mayaku = "mayaku.cli.__main__:app"

[project.urls]
Homepage = "https://github.com/datamarkin/mayaku"

[tool.hatch.build.targets.wheel]
packages = ["src/mayaku"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E", "F", "W",   # pycodestyle / pyflakes
  "I",             # isort
  "B",             # bugbear
  "UP",            # pyupgrade
  "SIM",           # simplify
  "RUF",
]
ignore = [
  "E501",  # line length handled by formatter
  # The codebase uses real Unicode glyphs (× for multiplication, – for
  # en-dash) in prose docstrings and comments — they read better than
  # ASCII substitutes. We don't put them in identifiers, so the
  # ambiguity ruff is guarding against doesn't apply.
  "RUF002", "RUF003",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "SIM"]
# Typer's idiom puts function calls in argument defaults; that's the
# whole point of `typer.Argument(...)` / `typer.Option(...)` and
# matches the official examples.
"src/mayaku/cli/__main__.py" = ["B008"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_reexport = true
disallow_untyped_decorators = false  # torch decorators are untyped
files = ["src/mayaku"]

[[tool.mypy.overrides]]
module = [
  "torchvision.*",
  "pycocotools.*",
  "coremltools.*",
  "openvino.*",
  "onnxruntime.*",
  "tensorrt.*",
]
ignore_missing_imports = true

# torch.amp re-exports its public surface via plain `from .x import y`
# (no `as y`), which trips mypy's no_implicit_reexport. Allow it for torch.
[[tool.mypy.overrides]]
module = ["torch", "torch.*"]
implicit_reexport = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
  "cuda: requires a CUDA device",
  "mps: requires an Apple-Silicon MPS device",
  "multi_gpu: requires >= 2 CUDA devices",
  "onnx: requires onnx package + onnxruntime or onnxruntime-gpu",
  "coreml: requires coremltools (macOS)",
  "openvino: requires openvino runtime",
  "tensorrt: requires a CUDA device + tensorrt runtime",
  "slow: long-running test",
]
filterwarnings = [
  # pycocotools' __array__ doesn't accept numpy 2's `copy=` kwarg.
  # Correctness unaffected — the fallback still produces the right
  # bitmap. Scoped to the pycocotools module so any genuine numpy-2
  # warning from our own code still surfaces. Upstream fix tracked in
  # cocodataset/cocoapi.
  "ignore:.*__array__ implementation.*:DeprecationWarning:pycocotools.*",
  # The LR-scheduler unit tests in test_engine_optim.py exercise the
  # schedule math without an actual optimizer.step(), which trips
  # PyTorch's "scheduler.step() before optimizer.step()" warning.
  # The runtime sequencing is correct in the trainer (LRScheduler
  # hook fires only after_step), so the warning is a test-shape
  # artifact, not a real ordering bug.
  "ignore:Detected call of:UserWarning",
  # We pin `dynamo=False` in ONNXExporter to avoid the new exporter's
  # `onnxscript` dep; PyTorch warns that the legacy path will go
  # away "starting in PyTorch 2.9". Revisit when we adopt the dynamo
  # path (will need `onnxscript` in `[onnx]`).
  "ignore:You are using the legacy TorchScript-based ONNX export:DeprecationWarning",
  "ignore:The feature will be removed:DeprecationWarning:torch.onnx.*",
  # CoreML conversion still requires `torch.jit.trace`; PT 2.x warns
  # that JIT will be removed eventually. Revisit when coremltools
  # supports torch.export-based traces.
  "ignore:`torch.jit.trace.*` is deprecated:DeprecationWarning",
]
